# openapi.yml
openapi: 3.0.3
info:
  title: Yalla Backend API
  version: 1.0.0
  description: Versioned API contract for Yalla clients and dashboard.
paths:
...
  /api/v1/courier/orders/{order_id}/status/:
    patch:
      operationId: v1_courier_orders_status_partial_update_by_order_id
      parameters:
      - in: path
        name: order_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/orders/:
    get:
      operationId: v1_orders_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderListList'
          description: ''
    post:
      operationId: v1_orders_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderCreate'
          description: ''
...
  /api/v1/orders/{order_id}/status/:
    patch:
      operationId: v1_orders_status_partial_update_by_order_id
      parameters:
      - in: path
        name: order_id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v2/courier/orders/{order_id}/status/:
    patch:
      operationId: v2_courier_orders_status_partial_update_by_order_id
      parameters:
      - in: path
        name: order_id
        schema:
          type: integer
        required: true
      tags:
      - v2
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v2/orders/:
    get:
      operationId: v2_orders_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - v2
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderListList'
          description: ''
    post:
      operationId: v2_orders_create
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminOrderCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderCreate'
          description: ''
...
  /api/v2/orders/{order_id}/status/:
    patch:
      operationId: v2_orders_status_partial_update_by_order_id
      parameters:
      - in: path
        name: order_id
        schema:
          type: integer
        required: true
      tags:
      - v2
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    AdminOrderCreate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: integer
        customer:
          type: object
          additionalProperties: {}
          readOnly: true
        delivery_address_id:
          type: integer
          nullable: true
        delivery_address:
          type: object
          additionalProperties: {}
          nullable: true
          readOnly: true
        shipping_company_id:
          type: integer
          readOnly: true
        shipping_company:
          allOf:
          - $ref: '#/components/schemas/ShippingCompanySummary'
          readOnly: true
        assigned_representative_id:
          type: integer
          readOnly: true
        assigned_representative:
          allOf:
          - $ref: '#/components/schemas/AssignedRepresentativeSummary'
          readOnly: true
        market_id:
          type: integer
        market:
          allOf:
          - $ref: '#/components/schemas/MarketSummary'
          readOnly: true
        order_scope:
          readOnly: true
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/OrderScopeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        service_city_id:
          type: integer
          nullable: true
        service_city:
          allOf:
          - $ref: '#/components/schemas/ServiceCitySummary'
          readOnly: true
        delivery_area_id:
          type: integer
          readOnly: true
          nullable: true
        delivery_area:
          allOf:
          - $ref: '#/components/schemas/DeliveryAreaSummary'
          readOnly: true
        delivery_type:
          allOf:
          - $ref: '#/components/schemas/DeliveryTypeEnum'
          readOnly: true
        fulfillment_type:
          allOf:
          - $ref: '#/components/schemas/FulfillmentTypeEnum'
          readOnly: true
        external_shipping_status:
          allOf:
          - $ref: '#/components/schemas/ExternalShippingStatusEnum'
          readOnly: true
        delivery_price_status:
          type: string
          readOnly: true
        eta_min_minutes:
          type: integer
          readOnly: true
          nullable: true
        eta_max_minutes:
          type: integer
          readOnly: true
          nullable: true
        payment_method:
          type: string
          maxLength: 50
        discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        description:
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/OrderStatusEnum'
          readOnly: true
        review_status:
          allOf:
          - $ref: '#/components/schemas/ReviewStatusEnum'
          readOnly: true
        delivery_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        subtotal_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        multi_market_fee_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          readOnly: true
        multi_market_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        total_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        image:
          type: string
          nullable: true
          readOnly: true
        assigned_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        delivered_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        delivery_note:
          type: string
        delivery_proof:
          type: string
          nullable: true
          readOnly: true
        approved_by:
          type: integer
          readOnly: true
          nullable: true
        approved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        rejected_by:
          type: integer
          readOnly: true
          nullable: true
        rejected_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        rejection_reason:
          type: string
          readOnly: true
        is_multi_market:
          type: boolean
          readOnly: true
        market_count:
          type: integer
          readOnly: true
        market_names_summary:
          type: string
          readOnly: true
        market_sections:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        grouped_items:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        grouped_offers:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        pickup_stops:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        history:
          type: array
          items:
            $ref: '#/components/schemas/OrderEvent'
          readOnly: true
        allowed_statuses:
          type: array
          items:
            type: string
          readOnly: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderItemCreate'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderOfferCreate'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - allowed_statuses
      - approved_at
      - approved_by
      - assigned_at
      - assigned_representative
      - assigned_representative_id
      - created_at
      - customer
      - delivered_at
      - delivery_address
      - delivery_area
      - delivery_area_id
      - delivery_price
      - delivery_price_status
      - delivery_proof
      - delivery_type
      - discount
      - eta_max_minutes
      - eta_min_minutes
      - external_shipping_status
      - fulfillment_type
      - grouped_items
      - grouped_offers
      - history
      - id
      - image
      - is_multi_market
      - market
      - market_count
      - market_names_summary
      - market_sections
      - multi_market_fee
      - multi_market_fee_rate
      - order_scope
      - payment_method
      - pickup_stops
      - rejected_at
      - rejected_by
      - rejection_reason
      - review_status
      - service_city
      - shipping_company
      - shipping_company_id
      - status
      - subtotal_price
      - total_price
      - updated_at
      - user_id
    AdminOrderCreateRequest:
      type: object
      properties:
        user_id:
          type: integer
        delivery_address_id:
          type: integer
          nullable: true
        market_id:
          type: integer
        service_city_id:
          type: integer
          nullable: true
        payment_method:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string
        delivery_note:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderItemCreateRequest'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/AdminOrderOfferCreateRequest'
        market_order:
          type: array
          items:
            type: integer
            minimum: 1
          writeOnly: true
      required:
      - payment_method
      - user_id
...
    AdminOrderItemCreateRequest:
      type: object
      properties:
        variant_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
        addition_ids:
          type: array
          items:
            type: integer
            minimum: 1
      required:
      - quantity
      - variant_id
...
    AdminOrderOfferCreateRequest:
      type: object
      properties:
        offer_id:
          type: integer
      required:
      - offer_id
    AssignedRepresentativeSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        phone:
          type: string
          nullable: true
          maxLength: 30
        avatar:
          type: string
          nullable: true
          readOnly: true
        avatar_url:
          type: string
          nullable: true
          readOnly: true
        service_city_id:
          type: integer
          readOnly: true
        service_city:
          type: object
          additionalProperties: {}
          nullable: true
          readOnly: true
        is_available:
          type: boolean
          readOnly: true
        vehicle_type:
          type: string
          readOnly: true
        plate_number:
          type: string
          readOnly: true
      required:
      - avatar
      - avatar_url
      - id
      - is_available
      - name
      - plate_number
      - service_city
      - service_city_id
      - vehicle_type
...
    BlankEnum:
      enum:
      - ''
...
    DeliveryAreaSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        service_city_id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        delivery_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,2})?$
        is_active:
          type: boolean
      required:
      - delivery_price
      - id
      - name
      - service_city_id
...
    DeliveryTypeEnum:
      enum:
      - fixed_area
      - delivery
      type: string
      description: |-
        * `fixed_area` - Fixed area
        * `delivery` - Delivery
...
    ExternalShippingStatusEnum:
      enum:
      - not_required
      - pending_quote
      - awaiting_customer_approval
      - quoted
      type: string
      description: |-
        * `not_required` - Not required
        * `pending_quote` - Pending quote
        * `awaiting_customer_approval` - Awaiting customer approval
        * `quoted` - Quoted
    FulfillmentTypeEnum:
      enum:
      - direct
      - external_shipping
      type: string
      description: |-
        * `direct` - Direct delivery
        * `external_shipping` - External shipping
...
    MarketSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        branch:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/MarketSummaryStatusEnum'
      required:
      - id
      - name
...
    MarketSummaryStatusEnum:
      enum:
      - active
      - inactive
      type: string
      description: |-
        * `active` - Active
        * `inactive` - Inactive
...
    NullEnum:
      enum:
      - null
...
    OrderList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        customer:
          type: object
          additionalProperties: {}
          readOnly: true
        status:
          $ref: '#/components/schemas/OrderStatusEnum'
        review_status:
          $ref: '#/components/schemas/ReviewStatusEnum'
        order_scope:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/OrderScopeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        market:
          allOf:
          - $ref: '#/components/schemas/MarketSummary'
          readOnly: true
        market_count:
          type: integer
          readOnly: true
        market_names_summary:
          type: string
          readOnly: true
        has_offer:
          type: boolean
          readOnly: true
        offer_titles:
          type: array
          items:
            type: string
          readOnly: true
        delivery_address:
          type: object
          additionalProperties: {}
          nullable: true
          readOnly: true
        delivery_type:
          $ref: '#/components/schemas/DeliveryTypeEnum'
        fulfillment_type:
          $ref: '#/components/schemas/FulfillmentTypeEnum'
        external_shipping_status:
          $ref: '#/components/schemas/ExternalShippingStatusEnum'
        delivery_price_status:
          type: string
          readOnly: true
        delivery_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        subtotal_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        multi_market_fee_rate:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        multi_market_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        total_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        assigned_representative_id:
          type: integer
          nullable: true
          readOnly: true
        assigned_representative:
          allOf:
          - $ref: '#/components/schemas/AssignedRepresentativeSummary'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - assigned_representative
      - assigned_representative_id
      - created_at
      - customer
      - delivery_address
      - delivery_price_status
      - has_offer
      - id
      - market
      - market_count
      - market_names_summary
      - offer_titles
      - updated_at
...
    OrderScopeEnum:
      enum:
      - general
      - service_city
      type: string
      description: |-
        * `general` - General
        * `service_city` - Service city
    OrderStatusEnum:
      enum:
      - pending
      - confirmed
      - assigned
      - picked_up
      - delivered
      - failed_delivery
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `confirmed` - Confirmed
        * `assigned` - Assigned
        * `picked_up` - Picked Up
        * `delivered` - Delivered
        * `failed_delivery` - Failed Delivery
        * `cancelled` - Cancelled
...
    PaginatedOrderListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/OrderList'
...
    ReviewStatusEnum:
      enum:
      - pending_review
      - approved
      - rejected
      type: string
      description: |-
        * `pending_review` - Pending Review
        * `approved` - Approved
        * `rejected` - Rejected
...
    ServiceCitySummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        delivery_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        is_active:
          type: boolean
      required:
      - id
      - name
...
    ShippingCompanySummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        logo_url:
          type: string
          nullable: true
          readOnly: true
      required:
      - id
      - logo_url
      - name
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT